move gpuinfo function
