+++ include/gpuinfo.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <string> 
+std::string gpuProgModel(std::string gpu); 
+
